update eflpp accordingly with the ecore_evas changes and fix a compilation error on 64 bits arch
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@35192 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/proto/eflpp/src/evas/eflpp_evas.cpp
@@ -527,7 +527,7 @@ void EvasObject::registerCallbacks()
 void EvasObject::dispatcher( void *data, Evas *evas, Evas_Object *evas_object, void *event_info )
 {
     EvasObject* object = static_cast<EvasObject*>( evas_object_data_get( evas_object, "obj_c++" ) );
-    int type = reinterpret_cast<int>(data);
+    long type = reinterpret_cast<long>(data);
 
     switch ( type )
     {
